Jwt token done and added a Weapon model

+++ Services/CharacterService/CharacterService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnetRpg.Data;
 using dotnetRpg.Dtos.Character;
 using dotnetRpg.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace dotnetRpg.Services.CharacterService
@@ -17,27 +19,34 @@ namespace dotnetRpg.Services.CharacterService
             new Character { Id = 1, Name="Sam"}
         };
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
         private readonly IMapper _mapper;
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
+            _httpContextAccessor = httpContextAccessor;
             _context = context;
             _mapper = mapper;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var ServiceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+            character.user = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
-            ServiceResponse.Data =await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            ServiceResponse.Data = await _context.Characters
+            .Where (c => c.user.Id == GetUserId())
+            .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
             return ServiceResponse;
         }
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var ServiceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _context.Characters.ToListAsync();
+            var dbCharacters = await _context.Characters.Where(c => c.user.Id == GetUserId()).ToListAsync();
             ServiceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return ServiceResponse;
         }
@@ -56,8 +65,12 @@ namespace dotnetRpg.Services.CharacterService
 
             try
             {
-                Character character =await _context.Characters.FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
+                Character character = await _context.Characters
+                .Include(c => c.user)
+                .FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
 
+                if(character.user.Id == GetUserId())
+                {
                 character.Name = updateCharacter.Name;
                 character.HitPoints = updateCharacter.HitPoints;
                 character.Strenght = updateCharacter.Strenght;
@@ -68,6 +81,12 @@ namespace dotnetRpg.Services.CharacterService
                 await _context.SaveChangesAsync();
 
                 serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else
+                {
+                    serviceResponse.Success =false;
+                    serviceResponse.Message ="Character not found";
+                }
             }
             catch (Exception ex)
             {
@@ -85,10 +104,21 @@ namespace dotnetRpg.Services.CharacterService
 
             try
             {
-                Character character =await _context.Characters.FirstAsync(c => c.Id == id);
+                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.user.Id ==GetUserId());
+                if(character != null)
+                {
                 _context.Characters.Remove(character);
                 await _context.SaveChangesAsync();
-                serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+
+                serviceResponse.Data = _context.Characters
+                .Where(c =>c.user.Id == GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success =false;
+                    serviceResponse.Message ="Character not found";
+                }
             }
             catch (Exception ex)
             {
@@ -98,5 +128,6 @@ namespace dotnetRpg.Services.CharacterService
 
             return serviceResponse;
         }
+
     }
 }
